+++ src/components/compound/list-view.tsx
@@ -4,7 +4,7 @@ interface ListViewProps {
   items: Array<any> | undefined | null;
   renderItem: (item: any, index: number) => React.ReactNode;
   placeholder?: React.ReactNode;
-  noItemsElement?: string;
+  noItemsElement?: React.ReactNode;
   footerElement?: React.ReactNode;
   className?: string;
   loading?: boolean;
@@ -16,7 +16,7 @@ interface ListViewProps {
  * @param {Array<any>} items - The list of items to be rendered
  * @param {React.ReactNode} renderItem - The function that renders each item
  * @param {React.ReactNode} placeholder - The placeholder to be displayed when the list is empty
- * @param {string} noItemsElement - The message to be displayed when the list is empty
+ * @param {React.ReactNode} noItemsElement - The message to be displayed when the list is empty
  * @param {React.ReactNode} footerElement - The element to be displayed at the bottom of the list
  * @param {string} className - The class name to be applied to the list view
  * @returns {React.ReactNode} - The list view component
@@ -36,7 +36,11 @@ export const ListView: React.FC<ListViewProps> = ({
   }
 
   if (!items || items === null || items.length === 0) {
-    return <>{noItemsElement && noItemsElement}</>;
+    if (noItemsElement) {
+      return <>{noItemsElement}</>;
+    } else {
+      return null;
+    }
   }
 
   return (
